\name{genDataset}
\alias{genDataset}
\title{genDataset}
\usage{
  genDataset(nC,popFreq)
}
\arguments{
  \item{nC}{Number of contributors in model.}

  \item{popFreq}{A list of allele frequencies for a given
  population.}
}
\description{
  genDataset samples a random mixture with gaussian peak
  height.
}
\details{
  Simple function for generating a random peak height
  mixture.
}
\author{
  Oyvind Bleka <Oyvind.Bleka.at.fhi.no>
}

