\name{mastermix-package}
\docType{package}
\title{
Procedure for deconvolving STR DNA mixtures}
\description{
The procedure sub-optimizes the mixture proportion simultaneous
    with combined genotypes by assuming the STR response variation as normal
    distributed.}
\details{
\tabular{ll}{
Package: \tab mastermix\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-04-18\cr
License: \tab  GPL(>=2) \cr
}
}
\author{
Oyvind Bleka <Oyvind.bleka@fhi.no>
Maintainer: Oyvind Bleka <Oyvind.bleka@fhi.no>
}
\references{
Tvedebrink,T, et.al.(2012). Identifying contributors of
  DNA mixtures by means of quantitative information of STR
  typing. Journal of Computational Biology, 19(7),887-902.
}

