\name{mastermixTK}
\alias{mastermixTK}
\title{mastermixTK}
\usage{
mastermixTK(envirfile=NULL)
}
\arguments{
  \item{envirfile}{A optional file to previous saved project of the program}
}
\description{
Software to evaluate weight-of-evidence for both qualitative and continuous models for STR DNA mixtures in addition to do deconvolution and database searching. 
}
\details{
mastermixTK opens the GUI "Mixture Tool" which performs weight-of-evidence for both qualitative and continuous models for STR DNA mixtures. It also provides extensive tools for deconvolution and database searching. See ?mastermix for more information.
}
\author{
Oyvind Bleka <Oyvind.Bleka.at.fhi.no>
}

\keyword{database search,}
\keyword{deconvolution,}
\keyword{Weight of evidence}

