\name{plotEPG}
\alias{plotEPG}
\title{plotEPG}
\usage{
plotEPG(Data, kitname, sname = "")
}
\arguments{
  \item{Data}{List of adata- and hdata-elements.}

  \item{kitname}{name of kit.}

  \item{sname}{label text.}
}
\description{
EPG plotter maintained by Oskar Hansson.
}
\details{
Plots peak height with corresponding allele for a sample.
}
\author{
Oyvind Bleka <Oyvind.Bleka.at.fhi.no>
}

